+++ tests/handler_test.py
@@ -5,6 +5,8 @@ import os
 from ipinfo.cache.default import DefaultCache
 from ipinfo.details import Details
 from ipinfo.handler import Handler
+from ipinfo import handler_utils
+import ipinfo
 import pytest
 
 
@@ -19,15 +21,14 @@ def test_init():
 def test_headers():
     token = "mytesttoken"
     handler = Handler(token)
-    headers = handler._get_headers()
+    headers = handler_utils.get_headers(token)
 
     assert "user-agent" in headers
     assert "accept" in headers
     assert "authorization" in headers
 
 
-@pytest.mark.parametrize("n", range(5))
-def test_get_details(n):
+def test_get_details():
     token = os.environ.get("IPINFO_TOKEN", "")
     handler = Handler(token)
     details = handler.getDetails("8.8.8.8")
@@ -75,25 +76,35 @@ def test_get_details(n):
 
         domains = details.domains
         assert domains["ip"] == "8.8.8.8"
-        assert domains["total"] == 12988
+        # NOTE: actual number changes too much
+        assert "total" in domains
         assert len(domains["domains"]) == 5
 
 
-@pytest.mark.parametrize("n", range(5))
-def test_get_batch_details(n):
+#############
+# BATCH TESTS
+#############
+
+_batch_ip_addrs = ["1.1.1.1", "8.8.8.8", "9.9.9.9"]
+
+
+def _prepare_batch_test():
+    """Helper for preparing batch test cases."""
     token = os.environ.get("IPINFO_TOKEN", "")
     if not token:
         pytest.skip("token required for batch tests")
     handler = Handler(token)
-    ips = ["1.1.1.1", "8.8.8.8", "9.9.9.9"]
-    details = handler.getBatchDetails(ips)
+    return handler, token, _batch_ip_addrs
 
+
+def _check_batch_details(ips, details, token):
+    """Helper for batch tests."""
     for ip in ips:
         assert ip in details
         d = details[ip]
         assert d["ip"] == ip
-        assert d["country"] == "US"
-        assert d["country_name"] == "United States"
+        assert "country" in d
+        assert "country_name" in d
         if token:
             assert "asn" in d
             assert "company" in d
@@ -102,52 +113,17 @@ def test_get_batch_details(n):
             assert "domains" in d
 
 
-def test_builtin_ip_types():
-    handler = Handler()
-    fake_details = {"country": "US", "ip": "127.0.0.1", "loc": "12.34,56.78"}
-
-    handler._requestDetails = lambda x: fake_details
-
-    details = handler.getDetails(IPv4Address(fake_details["ip"]))
-    assert isinstance(details, Details)
-    assert details.country == fake_details["country"]
-    assert details.country_name == "United States"
-    assert details.ip == fake_details["ip"]
-    assert details.loc == fake_details["loc"]
-    assert details.longitude == "56.78"
-    assert details.latitude == "12.34"
+@pytest.mark.parametrize("batch_size", [None, 1, 2, 3])
+def test_get_batch_details(batch_size):
+    handler, token, ips = _prepare_batch_test()
+    details = handler.getBatchDetails(ips, batch_size=batch_size)
+    _check_batch_details(ips, details, token)
 
 
-def test_json_serialization():
-    handler = Handler()
-    fake_details = {
-        "asn": {
-            "asn": "AS20001",
-            "domain": "twcable.com",
-            "name": "Time Warner Cable Internet LLC",
-            "route": "104.172.0.0/14",
-            "type": "isp",
-        },
-        "city": "Los Angeles",
-        "company": {
-            "domain": "twcable.com",
-            "name": "Time Warner Cable Internet LLC",
-            "type": "isp",
-        },
-        "country": "US",
-        "country_name": "United States",
-        "hostname": "cpe-104-175-221-247.socal.res.rr.com",
-        "ip": "104.175.221.247",
-        "loc": "34.0293,-118.3570",
-        "latitude": "34.0293",
-        "longitude": "-118.3570",
-        "phone": "323",
-        "postal": "90016",
-        "region": "California",
-    }
-
-    handler._requestDetails = lambda x: fake_details
-
-    details = handler.getDetails(fake_details["ip"])
-    assert isinstance(details, Details)
-    assert json.dumps(details.all)
+@pytest.mark.parametrize("batch_size", [1, 2])
+def test_get_batch_details_total_timeout(batch_size):
+    handler, token, ips = _prepare_batch_test()
+    with pytest.raises(ipinfo.exceptions.TimeoutExceededError):
+        handler.getBatchDetails(
+            ips, batch_size=batch_size, timeout_total=0.001
+        )